refactor: update migration schemas and improve documentation for vendor recommendations

+++ API/database/migrations/2026_03_11_000001_create_users_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->string('locale')->default('en');
             $table->timestamp('email_verified_at')->nullable();
             $table->timestamp('last_login_at')->nullable();
-            $table->integer('failed_login_attempts')->default(0);
+            $table->unsignedSmallInteger('failed_login_attempts')->default(0);
             $table->string('lockout_reason')->nullable();
             $table->timestamp('lockout_expires_at')->nullable();
             $table->boolean('mfa_enabled')->default(false);
